integrate the ProjectManager functionality directly into GeoparserV2
